Moved the Card, HandResult and PokerHand classes into the core package since they are related to the game logic
