+++ src/main/java/javatro_core/Card.java
@@ -1,4 +1,4 @@
-package Javatro;
+package javatro_core;
 
 public record Card(Rank rank, Suit suit) {
     public enum Rank {
